feat: add input password function and submit event
